Disable equipping items and looting chests while in an active adventure. This should alleviate issue #54 and #55
enable clearing of active adventures inside a server - This *should* never be required although in practice it occurs every so often requiring the cog to be reloaded. (This should alleviate #60)

Add ability to clear character sheets and remove items from users. Solves #58

Found better solution for #52 and solve #57.

Allow items to be given with negative stats. Solves #56